Add main package structure
Now the main source code can be found as part of the same package
`wentrivia`, and is divided in `trivia.py`, containing the logic for the
game, and `trivia_cog.py`, containing an implementation for said game as
a discord.py cog.

The versioning isn't important, but from now on, the format will be
`major version.minor version`, there won't be a number for bug fixes.

Other minimal changes:
- Added a `pathlib.Path` instance to load the question files to avoid
having issues with relatives path depending on the working
directory, e.g., running the program several directory levels up.)
- Added `__main__.py`. Now the testing example should be run with `python
-m wentrivia`. The TRIVIA_BOT_TOKEN should be set for this to work.
- Not setting now prints a message instead of raising an error.