Add possibility to specify the number of loaded questions
Now a specific number of questions can be loaded, following a rule: as
long as there are enough questions to hit the desired number, the program won't load the same
ones more than once. For example, if there are 50 questions and the
user wants to load a pool of 30, all 30 questions will be unique.

This doesn't try to have a low number of repeated questions beyond
that: if there are only 5 questions and the user wants a pool 30, the
first 5 to be loaded only once, but the remaining 25 are going to be
randomly selected based on the same 5 without any specific
criteria. That means the same question could be loaded another 25
times, or the 5 questions might be found in a more equitable
distribution, but just by chance, not design.