Add abstract classes and mixins to make trivia games
The new `ABCTrivia` class allows to quickly customize and make new
trivia games, with different rules and behaviours, by combining
combining mixins.

The class itself doesn't contain much logic, but sets a few rules to
(ideally) follow:

- `load_questions` should load the questions and store them in
`self.questions_pool`
- `check_answer` checks if an answer is correct. The arguments are a
`message` and a `question`, which can be of any type when overriden.
- `start` starts the game. The main logic doesn't need to be here, and
each implementation should be free to choose where that goes. `start`
only asks for `*args` and `**kwargs`, which could be passed to a
different function.
- `self.factory` should be set to a factory for the type of
`Question`, a dataclass that describes them and can be extended when
needed.

Some basic mixins were implemented for this commit:

- `JSONLoaderMixin`: loads `questions`
- `ForgivingCheckerMixin`: sets the factory callable to `ForgivingQuestion`
- `RegularLogicMixin`: defines the rules for a competitive, classic
- trivia
game.

These are used as follows:

```python
class ForgivingTrivia(RegularLogicMixin, ForgivingCheckerMixin, JSONLoaderMixin):
    def __init__(self, ctx: Context):
        super().__init__(ctx=ctx)
```